requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.40
Flask-Login==0.5.0
Flask-Bcrypt==1.0.1
bcrypt==3.1.7
Flask-WTF==0.15.1
WTForms==2.3.3
email_validator==1.1.3
PyMySQL==1.0.2
python-dotenv==0.19.2
waitress==2.0.0
jinja2==3.0.3
itsdangerous==2.0.1
click==8.0.4
importlib-metadata==4.8.3
typing-extensions==4.1.1
zipp==3.6.0
